requirements.txt
sanic==22.6.2
transformers==4.24.0
accelerate==0.14.0 
sentencepiece==0.1.97 
onnx==1.13.1
onnxruntime==1.14.1


accelerate==0.14.0
aiofiles==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
coloredlogs==15.0.1
filelock==3.10.0
flatbuffers==23.3.3
httptools==0.5.0
huggingface-hub==0.13.2
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
numpy==1.24.2
onnx==1.13.1
onnxruntime==1.14.1
packaging==23.0
Pillow==9.4.0
protobuf==3.20.3
psutil==5.9.4
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
sanic==22.6.2
sanic-routing==22.3.0
sentencepiece==0.1.97
sympy==1.11.1
tokenizers==0.13.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.24.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvloop==0.17.0
websockets==10.4
